requirements.txt
requests>=2.31.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
python-telegram-bot>=20.7
redis>=5.0.0
python-dotenv>=1.0.0
loguru>=0.7.2
aiohttp>=3.9.0
aiofiles>=23.2.1
fastapi>=0.109.0
uvicorn>=0.27.0
celery>=5.3.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
tenacity>=8.2.0
httpx>=0.26.0
lxml>=5.1.0
